+++ fedmsg/tests/test_crypto_gpg.py
@@ -19,54 +19,35 @@
 #
 import os
 
-try:
-    # For python-2.6, so we can do skipTest
-    import unittest2 as unittest
-except ImportError:
-    import unittest
+from nose.tools import raises
+import unittest
 
 import fedmsg.crypto
 
 SEP = os.path.sep
 here = SEP.join(__file__.split(SEP)[:-1])
 
-def skip_if_missing_libs(f):
-    def _wrapper(self, *args, **kw):
-        try:
-            import M2Crypto
-            import m2ext
-        except ImportError, e:
-            self.skipTest(str(e))
 
-        return f(self, *args, **kw)
-
-    return _wrapper
-
-class TestCrypto(unittest.TestCase):
+class TestCryptoGPG(unittest.TestCase):
 
     def setUp(self):
         self.config = {
-            # Normally this is /var/lib/puppet/ssl
-            'ssldir': SEP.join((here, 'test_certs/keys')),
-            # Normally this is 'app01.stg.phx2.fedoraproject.org'
-            'certname': 'shell-app01.phx2.fedoraproject.org',
-            'crl_location': "http://threebean.org/fedmsg-tests/crl.pem",
-            'crl_cache': "/tmp/crl.pem",
-            'crl_cache_expiry': 10,
-
+            'crypto_backend': 'gpg',
+            'ssldir': SEP.join((here, 'test_certs/keys_gpg')),
+            'certname': 'TODO -- fill me in.'
         }
 
     def tearDown(self):
         self.config = None
 
-    @skip_if_missing_libs
+    @raises(NotImplementedError)
     def test_full_circle(self):
         """ Try to sign and validate a message. """
         message = dict(msg='awesome')
         signed = fedmsg.crypto.sign(message, **self.config)
         assert fedmsg.crypto.validate(signed, **self.config)
 
-    @skip_if_missing_libs
+    @raises(NotImplementedError)
     def test_failed_validation(self):
         message = dict(msg='awesome')
         signed = fedmsg.crypto.sign(message, **self.config)
